+++ lang/attribute/program/src/declare_program/mod.rs
@@ -10,7 +10,7 @@ use common::gen_docs;
 use mods::{
     accounts::gen_accounts_mod, client::gen_client_mod, constants::gen_constants_mod,
     cpi::gen_cpi_mod, events::gen_events_mod, internal::gen_internal_mod, program::gen_program_mod,
-    types::gen_types_mod,
+    types::gen_types_mod, utils::gen_utils_mod,
 };
 
 pub struct DeclareProgram {
@@ -66,6 +66,9 @@ fn gen_program(idl: &Idl, name: &syn::Ident) -> proc_macro2::TokenStream {
     let client_mod = gen_client_mod(idl);
     let internal_mod = gen_internal_mod(idl);
 
+    // Utils
+    let utils_mod = gen_utils_mod(idl);
+
     quote! {
         #docs
         pub mod #name {
@@ -82,6 +85,8 @@ fn gen_program(idl: &Idl, name: &syn::Ident) -> proc_macro2::TokenStream {
             #cpi_mod
             #client_mod
             #internal_mod
+
+            #utils_mod
         }
     }
 }